src/use-case/checkIn/validate-check-in.ts
import { CheckIn } from '@prisma/client'
import { ResoucerNotFoundError } from '../err/resoucer-not-found'
import { CheckInRepository } from '@/repositories/checkin-repository'
import dayjs from 'dayjs'
import { LateCheckInValidateError } from '../err/late-checkin-validate-error'

interface ValidateCheckInUseCaseRequest {
  checkInId: string
}

interface ValidateCheckInUseCaseResponse {
  checkIn: CheckIn
}

export class ValidateCheckInUseCase {
  constructor(private checkInRepository: CheckInRepository) {}

  async execute({
    checkInId,
  }: ValidateCheckInUseCaseRequest): Promise<ValidateCheckInUseCaseResponse> {
    const checkIn = await this.checkInRepository.findById(checkInId)

    if (!checkIn) {
      throw new ResoucerNotFoundError()
    }

    const distanceInMinutesFromCreation = dayjs(new Date()).diff(
      checkIn.created_at,
      'minutes',
    )

    if (distanceInMinutesFromCreation > 20) {
      throw new LateCheckInValidateError()
    }

    checkIn.validated_at = new Date()

    await this.checkInRepository.save(checkIn)

    return { checkIn }
  }
}
